+++ src/calling/variants/preprocessing/haplotype_feature_index.rs
@@ -9,44 +9,7 @@ use rust_htslib::bcf::{
 };
 
 use crate::utils;
-
-#[derive(Hash, Debug, Eq, PartialEq)]
-pub(crate) enum HaplotypeIdentifier {
-    Event(Vec<u8>),
-    Phase { phase_set: u32, genotype: Genotype },
-}
-
-impl HaplotypeIdentifier {
-    pub(crate) fn from(record: &mut bcf::Record) -> Result<Option<Self>> {
-        if utils::is_bnd(record)? {
-            // TODO support records without EVENT tag.
-            if let Ok(Some(event)) = record.info(b"EVENT").string() {
-                let event = event[0];
-                return Ok(Some(HaplotypeIdentifier::Event(event.to_owned())));
-            }
-        } else {
-            if let Ok(ps_values) = record.format(b"PS").integer() {
-                let phase_set = ps_values[0][0];
-                if !phase_set.is_missing() {
-                    // phase set definition found
-                    if let Ok(genotypes) = record.genotypes() {
-                        let genotype = genotypes.get(0);
-                        if genotype.iter().any(|allele| match allele {
-                            GenotypeAllele::Phased(allele_idx) if *allele_idx > 0 => true,
-                            _ => false,
-                        }) {
-                            return Ok(Some(HaplotypeIdentifier::Phase {
-                                phase_set: phase_set as u32,
-                                genotype,
-                            }));
-                        }
-                    }
-                }
-            }
-        }
-        Ok(None)
-    }
-}
+use crate::variants::model::HaplotypeIdentifier;
 
 #[derive(Default, Debug)]
 pub(crate) struct HaplotypeFeatureIndex {
